[encointer-ceremonies] extract assignment_fn, generate_assignment_functions_params to encointer-ceremonies-assignments
