+++ ceremonies/assignment/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "encointer-ceremonies-assignment"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+
+# local deps
+encointer-primitives = { path = "../../primitives", default-features = false }
+
+# substrate deps
+sp-runtime = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate.git", default-features = false, branch = "master" }
+sp-std = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate.git", default-features = false, branch = "master" }
+
+
+[features]
+default = ["std"]
+std = [
+    "encointer-primitives/std",
+    "sp-std/std",
+    "sp-runtime/std",
+]
